+++ modules/jupyter-widget/src/notebook-utils.js
@@ -12,6 +12,29 @@ if (!window.require) {
   document.head.appendChild(requirejs);
 }
 
+export function loadCss(url) {
+  const link = document.createElement('link');
+  link.type = 'text/css';
+  link.rel = 'stylesheet';
+  link.href = url;
+  document.getElementsByTagName('head')[0].appendChild(link);
+}
+
+/**
+ * Hides a warning in the mapbox-gl.js library from surfacing in the notebook as text.
+ */
+export function hideMapboxCSSWarning() {
+  const missingCssWarning = document.getElementsByClassName('mapboxgl-missing-css')[0];
+  if (missingCssWarning) {
+    missingCssWarning.style.display = 'none';
+  }
+}
+
+export function updateDeck(inputJSON, {jsonConverter, deckgl}) {
+  const results = jsonConverter.convert(inputJSON);
+  deckgl.setProps(results);
+}
+
 export function initDeck({mapboxApiKey, container, jsonInput, tooltip, onComplete, handleClick}) {
   require(['mapbox-gl', 'h3', 's2Geometry'], mapboxgl => {
     require(['deck.gl', 'loaders.gl/csv'], (deck, loaders) => {
