+++ modules/jupyter-widget/src/plugin.js
@@ -4,7 +4,7 @@
  * https://github.com/jupyter-widgets/widget-ts-cookiecutter/blob/51e9fed8687e3b9cf1ed2fd307b7675e864f89ae/%7B%7Bcookiecutter.github_project_name%7D%7D/src/plugin.ts
  */
 import {IJupyterWidgetRegistry} from '@jupyter-widgets/base';
-import * as widgetExports from './lab-widget';
+import {DeckGLView, DeckGLModel} from './lab-widget';
 import {MODULE_NAME, MODULE_VERSION} from './version';
 
 const EXTENSION_ID = '@deck.gl/jupyter-widget:plugin';
@@ -21,6 +21,6 @@ function activateWidgetExtension(app, registry) {
   registry.registerWidget({
     name: MODULE_NAME,
     version: MODULE_VERSION,
-    exports: widgetExports
+    exports: {DeckGLView, DeckGLModel}
   });
 }
